+++ common/src/main/java/net/mehvahdjukaar/polytone/content/particle/custom/OverridingParticleFactory.java
@@ -12,9 +12,9 @@ import java.util.Map;
 
 public class OverridingParticleFactory<T extends ParticleOptions> implements ParticleProvider<T> {
 
-    private final CustomParticleFactory customFactory;
+    private final ICustomParticleFactory customFactory;
 
-    public OverridingParticleFactory(CustomParticleFactory factory) {
+    public OverridingParticleFactory(ICustomParticleFactory factory) {
         this.customFactory = factory;
     }